+++ src/starknet-docker-proxy.ts
@@ -28,9 +28,6 @@ export class StarknetDockerProxy extends DockerServer {
     }
 
     protected async getDockerArgs(): Promise<string[]> {
-        // get a valid free port
-        this.port = await this.getPort();
-
         // To access the files on host machine from inside the container, proper mounting has to be done.
         const volumes = ["-v", `${PROXY_SERVER_HOST_PATH}:${PROXY_SERVER_CONTAINER_PATH}`];
         volumes.push("-v", `${LEGACY_CLI_HOST_PATH}:${LEGACY_CLI_CONTAINER_PATH}`);
